+++ src/commands/mapshaper-cluster.js
@@ -47,7 +47,7 @@ internal.calcPolygonClusters = function(lyr, arcs, opts) {
   if (groupField && !lyr.data) stop("Missing attribute data table");
 
   // Populate mergeItems array
-  internal.findNeighbors(lyr.shapes, arcs).forEach(function(ab, i) {
+  internal.findPairsOfNeighbors(lyr.shapes, arcs).forEach(function(ab, i) {
     // ab: [a, b] indexes of two polygons
     var a = shapeItems[ab[0]],
         b = shapeItems[ab[1]],